pyproject.toml
[tool.poetry]
name = "ptp"
version = "1.0.0"
description = "PDF to PNG Conversion Service"
authors = ["Aliaksandr Skuratovich"]
license = "MIT"
packages = [
    { include = "app", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.109.0"
uvicorn = ">=0.27.1"
pdf2image = ">=1.17.0"
pydantic = ">=2.6.0"
python-multipart = ">=0.0.9"
mangum = ">=0.17.0"
pyyaml = ">=6.0.1"
poppler-utils = {version = "*", platform = "linux"}
magnum = "^19.0.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = ">=1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
